App refactor: Introduce StderrContextProvider
Improve App module's read- and compose-ability by moving StderrContext
concerns to a separate provider module.
